requirements.txt
antiorm==1.2.1
blinker==1.6.2
click==8.1.5
db==0.1.1
db-sqlite3==0.0.1
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
redis==5.0.1
Werkzeug==2.3.6
